# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PreventivoAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
...
  /documents/upload:
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a PDF or image document for price intelligence extraction
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedDocument"
...
  /documents/{id}/extract:
    post:
      operationId: extractDocument
      tags: [documents]
      summary: Trigger AI extraction on an uploaded document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Extraction complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedDocument"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete an uploaded document and its price intelligence data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    UploadUrlRequest:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType
    UploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
      required:
        - uploadURL
        - objectPath
...
    UploadedDocument:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
          nullable: true
        mimeType:
          type: string
        fileUrl:
          type: string
        status:
          type: string
          enum: [pending, processing, done, error]
        extractedData:
          $ref: "#/components/schemas/ExtractedDocumentData"
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - fileName
        - mimeType
        - fileUrl
        - status
        - createdAt
        - updatedAt
    ExtractedDocumentData:
      type: object
      properties:
        lavorazioni:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedLavorazione"
          nullable: true
        totale:
          type: number
          nullable: true
        zona:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
    ExtractedLavorazione:
      type: object
      properties:
        tipo:
          type: string
        prezzoUnitario:
          type: number
        um:
          type: string
          nullable: true
        zona:
          type: string
          nullable: true
      required:
        - tipo
        - prezzoUnitario
...